Support for Python3
Because of absolute imports, `uritemplates.py` has been merged with
`__init__.py`.
