+++ .travis.yml
@@ -3,6 +3,7 @@ python:
   - "2.5"
   - "2.6"
   - "2.7"
+  - "3.3"
   - "pypy"
 # dependencies
 install: pip install simplejson --use-mirrors
